% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_describe_backup_job}
\alias{backup_describe_backup_job}
\title{Returns metadata associated with creating a backup of a resource}
\usage{
backup_describe_backup_job(BackupJobId)
}
\arguments{
\item{BackupJobId}{[required] Uniquely identifies a request to AWS Backup to back up a resource.}
}
\description{
Returns metadata associated with creating a backup of a resource.
}
\section{Request syntax}{
\preformatted{svc$describe_backup_job(
  BackupJobId = "string"
)
}
}

\keyword{internal}
