% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_operations.R
\name{backup_list_protected_resources}
\alias{backup_list_protected_resources}
\title{Returns an array of resources successfully backed up by AWS Backup,
including the time the resource was saved, an Amazon Resource Name (ARN)
of the resource, and a resource type}
\usage{
backup_list_protected_resources(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The next item following a partial list of returned items. For example,
if a request is made to return \code{maxResults} number of items, \code{NextToken}
allows you to return more items in your list starting at the location
pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}
}
\description{
Returns an array of resources successfully backed up by AWS Backup,
including the time the resource was saved, an Amazon Resource Name (ARN)
of the resource, and a resource type.
}
\section{Request syntax}{
\preformatted{backup$list_protected_resources(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
