% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{filesystem_os}
\alias{filesystem_os}
\alias{filesystem}
\alias{is_filesystem}
\alias{is_filesystem_dir}
\alias{print.filesystem}
\alias{os_remove.filesystem}
\alias{os_remove_objects.filesystem}
\alias{os_write.filesystem}
\alias{os_update_tags.filesystem}
\alias{os_read.filesystem}
\alias{os_read_object.filesystem}
\alias{os_read_tags.filesystem}
\alias{os_exists.filesystem}
\alias{os_list.filesystem}
\alias{os_find.filesystem}
\title{Filesystem-based object store.}
\usage{
filesystem(path = getwd(), create = FALSE)

is_filesystem(x)

is_filesystem_dir(path, empty_ok = FALSE)

\method{print}{filesystem}(x, ...)

\method{os_remove}{filesystem}(store)

\method{os_remove_objects}{filesystem}(store, ids = os_list(store))

\method{os_write}{filesystem}(store, object, tags = list(),
  id = compute_id(object))

\method{os_update_tags}{filesystem}(store, id, tags)

\method{os_read}{filesystem}(store, id)

\method{os_read_object}{filesystem}(store, id)

\method{os_read_tags}{filesystem}(store, id)

\method{os_exists}{filesystem}(store, id)

\method{os_list}{filesystem}(store)

\method{os_find}{filesystem}(store, tags)
}
\arguments{
\item{path}{directory path; if does not exist and `create` is `TRUE`,
it will be created.}

\item{create}{whether to create `path` if does not exist.}

\item{x}{object to be tested.}

\item{empty_ok}{is an empty directory considered a valid filesystem
object store.}

\item{...}{Arguments passed on to \code{base::print}
\describe{
  \item{x}{an object used to select a method.}
}}

\item{store}{object store.}

\item{ids}{multiple object identifiers.}

\item{object}{R object to be written into object store.}

\item{tags}{a named `list` of tags for an object.}

\item{id}{object identifier; see [compute_id].}
}
\description{
Filesystem-based object store.
}
