% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{identifier}
\alias{identifier}
\alias{compute_id}
\alias{compute_id.default}
\alias{as_id}
\alias{print.identifier}
\alias{toString.identifier}
\alias{long}
\title{Object identifier.}
\usage{
compute_id(x)

\method{compute_id}{default}(x)

as_id(x)

\method{print}{identifier}(x, ...)

\method{toString}{identifier}(x, ...)

long(x)
}
\arguments{
\item{x}{object to compute identifier for, the identifier to print
or cast to `character` or a `character` value to cast as
identifier.}

\item{...}{Arguments passed on to \code{base::print}
\describe{
  \item{x}{an object used to select a method.}
}}
}
\value{
Object of class `"identifier"`.
}
\description{
`compute_id` calculates an unique identifier based on the
contents of the object `x`.

the default implementation of `compute_id` uses the
__sha1__ algorithm implemented in [digest::digest].

`as_id` turns a 40-character string into an identifier.

`long` forces printing of all 40 characters of an
identifier.
}
\examples{
i <- compute_id(iris)

toString(i) # returns 8-character string
toString(full(i)) returns 40-character string

print(i) # prints 8 characters
print(full(i)) # prints 40 characters
}
