% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memory_os}
\alias{memory_os}
\alias{memory}
\alias{is_memory}
\alias{print.memory}
\alias{os_remove.memory}
\alias{os_remove_objects.memory}
\alias{os_write.memory}
\alias{os_update_tags.memory}
\alias{os_read.memory}
\alias{os_read_object.memory}
\alias{os_read_tags.memory}
\alias{os_exists.memory}
\alias{os_list.memory}
\alias{os_find.memory}
\title{Memory-based object store.}
\usage{
memory()

is_memory(x)

\method{print}{memory}(x, ...)

\method{os_remove}{memory}(x)

\method{os_remove_objects}{memory}(x, ids = os_list(x))

\method{os_write}{memory}(store, object, tags = list(),
  id = compute_id(object))

\method{os_update_tags}{memory}(store, id, tags)

\method{os_read}{memory}(store, id)

\method{os_read_object}{memory}(store, id)

\method{os_read_tags}{memory}(store, id)

\method{os_exists}{memory}(store, id)

\method{os_list}{memory}(store)

\method{os_find}{memory}(store, tags)
}
\description{
Memory-based object store.
}
