% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mock.Storage.Broker.R
\name{Mock.Storage.Broker}
\alias{Mock.Storage.Broker}
\title{Mock Storage Broker}
\value{
A \code{list} of functions:
\itemize{
\item \code{Insert(entity, table)}
\item \code{Select(fields, table)}
\item \code{SelectWhereId(fields, table, id)}
\item \code{Update(entity, table)}
\item \code{Delete(id, table)}
}
}
\description{
Provide a mock data store with basic CRUD Operations.
}
