% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mock.Storage.Service.R
\name{Mock.Storage.Service}
\alias{Mock.Storage.Service}
\title{Mock Data Access Layer}
\value{
A \code{list} of utility functions:
\itemize{
\item \verb{[["Todo"]][["Insert"]](todo)}
\item \verb{[["Todo"]][["Select"]]()}
\item \verb{[["Todo"]][["SelectWhereId]](id)}
\item \verb{[["Todo"]][["Update"]](todo)}
\item \verb{[["Todo"]][["Delete"]](id)}
}
}
\description{
Typical operations perform on a Data Stores, for example a SQL Database, includes \code{Insert}, \code{Select}, \code{Update} and \code{Delete}.
This \code{Storage} package provides a simple function for each of these operations while hiding all unnessecary details.

\code{Mock.Storage.Service()} take a configuration as input and returns a list of functions, each representing a Data Store operation.
The list of functions is based on a simple Todo App example:
\itemize{
\item \code{Insert}, take a todo as input and adds the todo to the data store.
\item \code{Select}, returns a data frame contains all todo's in the data store.
\item \code{SelectWhereId}, take the Id and returns the macthing todo from the data store, if it exist.
\item \code{Update}, take a todo as input and update the task and status of the todo in the data store.
\item \code{Delete}, that the Id of a todo and removes that todo from the data store if it exist.
}

\code{Mock.Storage.Service()} also includes input validation and exception handling:
\itemize{
\item The \code{configuration} input is expected to be a list with \code{DSN}, \code{Username} and \code{Password}.
\item The \code{todo} input used with \code{Insert} and \code{Update} operations is expcted to be a data frame containing an \code{Id}, \code{Task} and \code{Status}
\item The \code{id} input used with \code{SelectWhereId} and \code{Delete} operations is expcted to not be null.
}
}
\examples{
  configurator <- Configuration.Service()
  configuration <- configurator[["GetODBCConfiguration"]]()

  operation <- configuration |> Mock.Storage.Service()

  todos <- operation[["Todo"]][["Select"]]()

  rm(configurator)
  rm(configuration)
  rm(operation)
  rm(todos)
}
