% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBC.Storage.Broker.R
\name{ODBC.Storage.Broker}
\alias{ODBC.Storage.Broker}
\title{ODBC Storage Broker}
\value{
A \code{list} of functions:
\itemize{
\item \code{ExecuteQuery(query)}
\item \code{Insert(entity, table)}
\item \code{Select(fields, table)}
\item \code{SelectWhereId(fields, table, id)}
\item \code{Update(entity, table)}
\item \code{Delete(id, table)}
}
}
\description{
Provide a ODBC based storage interface with basic CRUD Operations.
}
