% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_cors}
\alias{s3_delete_bucket_cors}
\title{Deletes the CORS configuration information set for the bucket}
\usage{
s3_delete_bucket_cors(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Deletes the CORS configuration information set for the bucket.
}
\section{Request syntax}{
\preformatted{s3$delete_bucket_cors(
  Bucket = "string"
)
}
}

\examples{
# The following example deletes CORS configuration on a bucket.
\donttest{s3$delete_bucket_cors(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
