% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_acl}
\alias{s3_get_bucket_acl}
\title{Gets the access control policy for the bucket}
\usage{
s3_get_bucket_acl(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Gets the access control policy for the bucket.
}
\section{Request syntax}{
\preformatted{svc$get_bucket_acl(
  Bucket = "string"
)
}
}

\keyword{internal}
