% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_encryption}
\alias{s3_get_bucket_encryption}
\title{Returns the server-side encryption configuration of a bucket}
\usage{
s3_get_bucket_encryption(Bucket)
}
\arguments{
\item{Bucket}{[required] The name of the bucket from which the server-side encryption
configuration is retrieved.}
}
\description{
Returns the server-side encryption configuration of a bucket.
}
\section{Request syntax}{
\preformatted{svc$get_bucket_encryption(
  Bucket = "string"
)
}
}

\keyword{internal}
