% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_location}
\alias{s3_get_bucket_location}
\title{Returns the region the bucket resides in}
\usage{
s3_get_bucket_location(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the region the bucket resides in.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_location(
  Bucket = "string"
)
}
}

\examples{
# The following example returns bucket location.
\donttest{s3$get_bucket_location(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
