% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_notification}
\alias{s3_get_bucket_notification}
\title{No longer used, see the GetBucketNotificationConfiguration operation}
\usage{
s3_get_bucket_notification(Bucket)
}
\arguments{
\item{Bucket}{[required] Name of the bucket to get the notification configuration for.}
}
\description{
No longer used, see the GetBucketNotificationConfiguration operation.
}
\section{Request syntax}{
\preformatted{svc$get_bucket_notification(
  Bucket = "string"
)
}
}

\examples{
# The following example returns notification configuration set on a
# bucket.
\donttest{svc$get_bucket_notification(
  Bucket = "examplebucket"
)}

# The following example returns notification configuration set on a
# bucket.
\donttest{svc$get_bucket_notification(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
