% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_policy_status}
\alias{s3_get_bucket_policy_status}
\title{Retrieves the policy status for an Amazon S3 bucket, indicating whether
the bucket is public}
\usage{
s3_get_bucket_policy_status(Bucket)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose policy status you want to
retrieve.}
}
\description{
Retrieves the policy status for an Amazon S3 bucket, indicating whether
the bucket is public.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_policy_status(
  Bucket = "string"
)
}
}

\keyword{internal}
