% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_request_payment}
\alias{s3_get_bucket_request_payment}
\title{Returns the request payment configuration of a bucket}
\usage{
s3_get_bucket_request_payment(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the request payment configuration of a bucket.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_request_payment(
  Bucket = "string"
)
}
}

\examples{
# The following example retrieves bucket versioning configuration.
\donttest{s3$get_bucket_request_payment(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
