% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_tagging}
\alias{s3_get_bucket_tagging}
\title{Returns the tag set associated with the bucket}
\usage{
s3_get_bucket_tagging(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the tag set associated with the bucket.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_tagging(
  Bucket = "string"
)
}
}

\examples{
# The following example returns tag set associated with a bucket
\donttest{s3$get_bucket_tagging(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
