% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_public_access_block}
\alias{s3_get_public_access_block}
\title{Retrieves the PublicAccessBlock configuration for an Amazon S3 bucket}
\usage{
s3_get_public_access_block(Bucket)
}
\arguments{
\item{Bucket}{[required] The name of the Amazon S3 bucket whose \code{PublicAccessBlock} configuration
you want to retrieve.}
}
\description{
Retrieves the \code{PublicAccessBlock} configuration for an Amazon S3 bucket.
}
\section{Request syntax}{
\preformatted{svc$get_public_access_block(
  Bucket = "string"
)
}
}

\keyword{internal}
