% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_list_bucket_metrics_configurations}
\alias{s3_list_bucket_metrics_configurations}
\title{Lists the metrics configurations for the bucket}
\usage{
s3_list_bucket_metrics_configurations(Bucket, ContinuationToken)
}
\arguments{
\item{Bucket}{[required] The name of the bucket containing the metrics configurations to
retrieve.}

\item{ContinuationToken}{The marker that is used to continue a metrics configuration listing that
has been truncated. Use the NextContinuationToken from a previously
truncated list response to continue the listing. The continuation token
is an opaque value that Amazon S3 understands.}
}
\description{
Lists the metrics configurations for the bucket.
}
\section{Request syntax}{
\preformatted{svc$list_bucket_metrics_configurations(
  Bucket = "string",
  ContinuationToken = "string"
)
}
}

\keyword{internal}
