% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_put_bucket_lifecycle}
\alias{s3_put_bucket_lifecycle}
\title{No longer used, see the PutBucketLifecycleConfiguration operation}
\usage{
s3_put_bucket_lifecycle(Bucket, ContentMD5, LifecycleConfiguration)
}
\arguments{
\item{Bucket}{[required]}

\item{ContentMD5}{}

\item{LifecycleConfiguration}{}
}
\description{
No longer used, see the PutBucketLifecycleConfiguration operation.
}
\section{Request syntax}{
\preformatted{svc$put_bucket_lifecycle(
  Bucket = "string",
  ContentMD5 = "string",
  LifecycleConfiguration = list(
    Rules = list(
      list(
        Expiration = list(
          Date = as.POSIXct(
            "2015-01-01"
          ),
          Days = 123,
          ExpiredObjectDeleteMarker = TRUE|FALSE
        ),
        ID = "string",
        Prefix = "string",
        Status = "Enabled"|"Disabled",
        Transition = list(
          Date = as.POSIXct(
            "2015-01-01"
          ),
          Days = 123,
          StorageClass = "GLACIER"|"STANDARD_IA"|"ONEZONE_IA"|"INTELLIGENT_TIERING"|"DEEP_ARCHIVE"
        ),
        NoncurrentVersionTransition = list(
          NoncurrentDays = 123,
          StorageClass = "GLACIER"|"STANDARD_IA"|"ONEZONE_IA"|"INTELLIGENT_TIERING"|"DEEP_ARCHIVE"
        ),
        NoncurrentVersionExpiration = list(
          NoncurrentDays = 123
        ),
        AbortIncompleteMultipartUpload = list(
          DaysAfterInitiation = 123
        )
      )
    )
  )
)
}
}

\keyword{internal}
