% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage.R
\name{Storage}
\alias{Storage}
\title{Storage}
\value{
A \code{list} of functions:
\itemize{
\item \code{Execute.Query(query)}
\item \code{Add(entity, table)}
\item \code{Retrieve(table, fields)}
\item \code{RetrieveWhereId(id, table, fields)}
\item \code{Modify(entity, table)}
\item \code{Remove(id, table)}
}
}
\description{
Provide an storage interface with basic CRUD Operations with ability to switch between MOCK or ODBC Storage Provider.
}
