% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODBC.Storage.Service.R
\name{ODBC.Storage.Service}
\alias{ODBC.Storage.Service}
\title{Storage Server}
\value{
A \code{list} of functions:
\itemize{
\item \code{ExecuteQuery(query)}
\item \code{Insert(entity, table)}
\item \code{Select(fields, table)}
\item \code{SelectWhereId(fields, table, id)}
\item \code{Update(entity, table)}
\item \code{Delete(id, table)}
}
}
\description{
Provide an storage interface with basic CRUD Operations with ability to switch between MOCK or ODBC Storage Provider.
}
