% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagegateway_operations.R
\name{storagegateway_describe_nfs_file_shares}
\alias{storagegateway_describe_nfs_file_shares}
\title{Gets a description for one or more Network File System (NFS) file shares
from a file gateway}
\usage{
storagegateway_describe_nfs_file_shares(FileShareARNList)
}
\arguments{
\item{FileShareARNList}{[required] An array containing the Amazon Resource Name (ARN) of each file share to
be described.}
}
\description{
Gets a description for one or more Network File System (NFS) file shares
from a file gateway. This operation is only supported for file gateways.
}
\section{Request syntax}{
\preformatted{svc$describe_nfs_file_shares(
  FileShareARNList = list(
    "string"
  )
)
}
}

\keyword{internal}
