% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Todo.Broker.R
\name{Todo.Broker}
\alias{Todo.Broker}
\title{Example Todo Broker}
\value{
A \code{list} of functions:
\itemize{
\item \code{insert(todo)}
\item \code{select(...)}
\item \code{select.where.Id(id)}
\item \code{update(todo)}
\item \code{delete(id)}
}
}
\description{
Provide a Broker which leverage the storage interface provided by this package.
The broker excludes structural and logical validation of data.
}
