% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Todos}
\alias{Todos}
\title{Todos Dataset}
\format{
A data frame with 3 rows and 5 columns:
\describe{
\item{id}{Unique identifier for the todo, stored as a UUID string.}
\item{task}{Description of the task.}
\item{status}{Status of the task.}
}
}
\source{
Generated synthetic data.
}
\usage{
Todos
}
\description{
This dataset contains todos for testing the package.
}
\keyword{datasets}
