% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{filesystem_os}
\alias{filesystem_os}
\alias{filesystem}
\alias{is_filesystem}
\alias{is_filesystem_dir}
\alias{print.filesystem}
\alias{os_remove.filesystem}
\alias{os_write.filesystem}
\alias{os_read.filesystem}
\alias{os_read_object.filesystem}
\alias{os_read_tags.filesystem}
\alias{os_exists.filesystem}
\alias{os_list.filesystem}
\alias{os_find.filesystem}
\title{Filesystem-based object store.}
\usage{
filesystem(path = getwd(), create = FALSE)

is_filesystem(x)

is_filesystem_dir(path, empty_ok = FALSE)

\method{print}{filesystem}(x, ...)

\method{os_remove}{filesystem}(x)

\method{os_write}{filesystem}(store, object, tags = list(),
  id = compute_id(object))

\method{os_read}{filesystem}(store, id)

\method{os_read_object}{filesystem}(store, id)

\method{os_read_tags}{filesystem}(store, id)

\method{os_exists}{filesystem}(store, id)

\method{os_list}{filesystem}(store)

\method{os_find}{filesystem}(store, lazy_tags)
}
\description{
Filesystem-based object store.
}
