% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R, R/memory.R
\name{toString.memory}
\alias{toString.memory}
\alias{memory_os}
\alias{memory}
\alias{is_memory}
\alias{print.memory}
\alias{os_remove.memory}
\alias{os_remove_objects.memory}
\alias{os_write.memory}
\alias{os_update_tags.memory}
\alias{os_read.memory}
\alias{os_read_object.memory}
\alias{os_read_tags.memory}
\alias{os_exists.memory}
\alias{os_list.memory}
\alias{os_find.memory}
\title{Memory-based object store.}
\usage{
\method{toString}{memory}(x, ...)

memory()

is_memory(x)

\method{print}{memory}(x, ...)

\method{toString}{memory}(x, ...)

\method{os_remove}{memory}(store)

\method{os_remove_objects}{memory}(store, ids = os_list(store))

\method{os_write}{memory}(store, object, tags = list(),
  id = compute_id(object))

\method{os_update_tags}{memory}(store, id, tags)

\method{os_read}{memory}(store, id)

\method{os_read_object}{memory}(store, id)

\method{os_read_tags}{memory}(store, id)

\method{os_exists}{memory}(store, id)

\method{os_list}{memory}(store)

\method{os_find}{memory}(store, tags)
}
\arguments{
\item{x}{object to be tested.}

\item{...}{Arguments passed on to \code{base::print}
\describe{
  \item{x}{an object used to select a method.}
}}

\item{store}{object store.}

\item{ids}{multiple object identifiers.}

\item{object}{R object to be written into object store.}

\item{tags}{a named `list` of tags for an object.}

\item{id}{object identifier; see [compute_id].}
}
\description{
Memory-based object store.
}
