% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{object_store}
\alias{is_object_store}
\alias{object_store}
\alias{os_find}
\alias{os_list}
\alias{os_read}
\alias{os_read_object}
\alias{os_read_tags}
\alias{os_write}
\title{Public interface for every object store.}
\usage{
is_object_store(x)

os_write(store, object, tags, id = hash(object))

os_read(store, id)

os_read_object(store, id)

os_read_tags(store, id)

os_find(storage, lazy_tags)

os_list(storage)
}
\description{
Basic operations:
\enumerate{
  \item store (write) an object with its tags
  \item restore (read) an object given its id
  \item find objects (ids) by tags
  \item extra utility calls
}
}

