% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{object_store}
\alias{object_store}
\alias{is_object_store}
\alias{os_remove}
\alias{os_remove_objects}
\alias{os_write}
\alias{os_update_tags}
\alias{os_read}
\alias{os_read_object}
\alias{os_read_tags}
\alias{os_exists}
\alias{os_find}
\alias{os_list}
\title{Public interface for every object store.}
\usage{
is_object_store(x)

os_remove(x)

os_remove_objects(x, ids = os_list(x))

os_write(store, object, tags, id = compute_id(object))

os_update_tags(store, id, tags)

os_read(store, id)

os_read_object(store, id)

os_read_tags(store, id)

os_exists(store, id)

os_find(storage, tags)

os_list(storage)
}
\value{
\code{os_write} returns object id.
}
\description{
Basic operations:
\enumerate{
  \item store (write) an object with its tags
  \item restore (read) an object given its id
  \item find objects (ids) by tags
  \item extra utility calls
}
}
