% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage.R
\name{Storage}
\alias{Storage}
\title{Storage}
\value{
A \code{list} of functions:
\itemize{
\item \code{execute.query(query)}
\item \code{add(entity, table)}
\item \code{retrieve(table, fields)}
\item \code{retrieve.where.id(id, table, fields)}
\item \code{modify(entity, table)}
\item \code{remove(id, table)}
}
}
\description{
Provide an storage interface with basic CRUD Operations with ability to switch between MOCK or ODBC Storage Provider.
}
