% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Todo.Service.R
\name{Todo.Service}
\alias{Todo.Service}
\title{Storage Server}
\value{
A \code{list} of functions:
\itemize{
\item \code{Insert(todo)}
\item \code{Select(...)}
\item \code{SelectWhereId(id)}
\item \code{Update(todo)}
\item \code{Delete(id)}
}
}
\description{
Provide an storage interface with basic CRUD Operations with ability to switch between MOCK or ODBC Storage Provider.
}
