% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Todo.Service.R
\name{Todo.Service}
\alias{Todo.Service}
\title{Example Todo Service}
\value{
A \code{list} of functions:
\itemize{
\item \code{add(todo)}
\item \code{retrieve(...)}
\item \code{retrieve.where.id(id)}
\item \code{modify(todo)}
\item \code{remove(id)}
}
}
\description{
Provide a Service which leverage the storage interface provided by this package.
The service includes structural and logical validation of data.
}
